Reorganize Python code structure
The code has been reorganized/mutualized.

- move resolvconf to hosts / resolv_conf plugins
- move according tests
- move utilities to root package
- remove modules/ package
