--- wazo_sysconfd/modules/tests/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2013,2015 Avencall
-# SPDX-License-Identifier: GPL-3.0-or-later
