+++ wazo_sysconfd/plugins/ha_config/tests/test_ha.py
