+++ wazo_sysconfd/plugins/hosts/plugin.py
@@ -1,13 +1,13 @@
 # Copyright 2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from wazo_sysconfd.modules import resolvconf
 from .http import router
+from .services import safe_init
 
 
 class Plugin:
     def load(self, dependencies: dict):
         api = dependencies['api']
         config = dependencies['config']
-        resolvconf.safe_init(config)
+        safe_init(config)
         api.include_router(router)
