+++ wazo_sysconfd/plugins/resolv_conf/services.py
@@ -6,16 +6,14 @@ import logging
 
 from time import time
 from shutil import copy2
-
-from wazo_sysconfd.exceptions import HttpReqError
-from wazo_sysconfd.modules.utilities import txtsubst
 from xivo import system
 
 from wazo_sysconfd import helpers
+from wazo_sysconfd.exceptions import HttpReqError
 from wazo_sysconfd.dns_lock import RESOLVCONFLOCK
+from wazo_sysconfd.utilities import txtsubst
 
 log = logging.getLogger('wazo_sysconfd.plugins.resolvconf')
-
 
 Rcc = {
     'hostname_file': os.path.join(os.path.sep, 'etc', 'hostname'),
